added styling and fixed bugs about fetching notes from backend and using access token

+++ front-end/src/components/NoteItem.jsx
@@ -1,16 +1,18 @@
 import React from "react";
+import "./NoteItem.css"
 
-const NoteItem = ({note, isActive, onSelect, onDelete}) => {
-
+const NoteItem = ({ note, isActive, onSelect, onDelete }) => {
     const formattedDate = note.updatedAt ? new Date(note.updatedAt).toLocaleString() : new Date(note.createdAt).toLocaleString();
 
     const handleDelete = (e) => {
         e.stopPropagation();
-        onDelete(note.id);
+        if (window.confirm("Are you sure you want to delete this note?")) {
+            onDelete(note.id);
+        }
     };
 
     return (
-        <div className={`noteItem ${isActive ? 'active': ''} `}
+        <div className={`noteItem ${isActive ? 'active' : ''}`}
              onClick={() => onSelect(note)}>
             <div className="noteItemHeader">
                 <h3>{note.title}</h3>
@@ -20,7 +22,7 @@ const NoteItem = ({note, isActive, onSelect, onDelete}) => {
             </div>
 
             <p className="notePreview">
-                {note.content.length > 50 ? note.content.substring(0,50) + '...' : note.content}
+                {note.content.length > 50 ? note.content.substring(0, 50) + '...' : note.content}
             </p>
             <p className="noteDate">{formattedDate}</p>
         </div>
